refactor: consolidate sorting and grouping utilities into reducerUtils
Since sortTools and getGroupLetter functions are only used in reducers,
moved them into reducerUtils.ts for better organization.

- Moved sortNumber, sortString, collator, and getGroupLetter to reducerUtils
- Updated all reducer imports to use reducerUtils
- Kept collator re-exported from sortTools for backward compatibility with SelectOptions.tsx
- Deleted getGroupLetter.ts as it's now in reducerUtils
