+++ src/components/CastAndCrew/CastAndCrew.reducer.ts
@@ -1,11 +1,12 @@
-import { getGroupLetter } from "~/utils/getGroupLetter";
 import {
   buildGroupValues,
   createNameFilter,
   type FilterableState,
   filterTools,
+  getGroupLetter,
+  sortNumber,
+  sortString,
 } from "~/utils/reducerUtils";
-import { sortNumber, sortString } from "~/utils/sortTools";
 
 import type { ListItemValue } from "./CastAndCrew";
