+++ src/components/Collections/Collections.reducer.ts
@@ -1,5 +1,9 @@
-import { createNameFilter, filterValues } from "~/utils/reducerUtils";
-import { sortNumber, sortString } from "~/utils/sortTools";
+import {
+  createNameFilter,
+  filterValues,
+  sortNumber,
+  sortString,
+} from "~/utils/reducerUtils";
 
 import type { ListItemValue } from "./Collections";
