--- src/utils/getGroupLetter.ts
@@ -1,17 +1,0 @@
-/**
- * Gets the group letter for a given string, typically used for alphabetical grouping.
- * Non-alphabetic characters are grouped under "#".
- *
- * @param str - The string to get the group letter from
- * @returns The uppercase first letter or "#" for non-alphabetic characters
- */
-export function getGroupLetter(str: string): string {
-  const letter = str.slice(0, 1);
-
-  // Check if the character is non-alphabetic (same in upper and lower case)
-  if (letter.toLowerCase() === letter.toUpperCase()) {
-    return "#";
-  }
-
-  return letter.toLocaleUpperCase();
-}
