+++ src/utils/sortTools.ts
@@ -1,21 +1,3 @@
-export const collator = new Intl.Collator("en", {
-  ignorePunctuation: true,
-  numeric: true,
-  sensitivity: "base",
-});
-
-export function sortNumber(a: number, b: number): number {
-  return a - b;
-}
-
-export function sortString(a: string, b: string): number {
-  if (a > b) {
-    return 1;
-  }
-
-  if (a < b) {
-    return -1;
-  }
-
-  return 0;
-}
+// Re-export collator from reducerUtils for backward compatibility
+// This is kept for SelectOptions.tsx which uses collator
+export { collator } from "./reducerUtils";
